address-space: split out address space region related code
Split out address space region related code into dedicated source file.
